# Conflicts:
#	Assets/Customizable Skybox/Customizable Skybox Documentaion.pdf.meta
#	Assets/Dreamteck/Splines/Examples.unitypackage.meta
#	Assets/Dreamteck/Splines/PlaymakerActions.unitypackage.meta
#	Assets/Dreamteck/Splines/TMPro.unitypackage.meta
#	Assets/Prefabs/Buildings/BuildManager.prefab.meta
#	Assets/Prefabs/Character/DemonManager.prefab.meta
#	Assets/Prefabs/Layer/LandLayerManager.prefab.meta
#	Assets/Prefabs/imprted/LowPolyMedievalPropsLite/LowPolyMedievalPropsLite_2020.3.0_HDRP_v1.0.unitypackage.meta
#	Assets/Prefabs/imprted/LowPolyMedievalPropsLite/LowPolyMedievalPropsLite_2020.3.0_URP_v1.0.unitypackage.meta

+++ Assets/Scripts/Fear/DemonFear.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 
 public class DemonFear : MonoBehaviour
 {
-    private int _fearLevel = 0;
-    public int FearLevel => _fearLevel;
+    private float _fearLevel = 0;
+    public float FearLevel => _fearLevel;
 
     private int _layer;
     public int Layer => _layer;
     [SerializeField] private int _layertHightDiff = 100;
 
-    public int DecayRate;
+    public float DecayRate;
 
 
 
@@ -21,18 +21,18 @@ public class DemonFear : MonoBehaviour
     }
 
     
-    public int GetFearLevel()
+    public float GetFearLevel()
     {
         return _fearLevel;
     }
   
 
-    public void IncreaseFear(int amount)
+    public void IncreaseFear(float amount)
     {
         _fearLevel += amount;
     }
 
-    public void DecreaseFear(int amount)
+    public void DecreaseFear(float amount)
     {
         _fearLevel -= amount;
     }
